refactor: split taxonomy.go into types + internal helpers (#540)
taxonomy.go was 534 lines — over the project's ~300-line guideline.
Split the precompute / deep-copy / category-derivation helpers out
into a new taxonomy_internal.go file. Pure refactor; no behaviour
change, no public API change.

Kept in taxonomy.go (321 lines):
- Public types: Fields, SensitivityConfig, SensitivityLabel,
  CategoryDef, EventDef, Taxonomy
- Public funcs: DevTaxonomy, EventDef.ResolvedSeverity, clampSeverity
- Const blocks: taxonomy version range, severity bounds

Moved to taxonomy_internal.go (229 lines):
- precomputeTaxonomy, deriveEventCategories, resolveEventSeverity,
  precomputeEventDef, sortedCopy
- deepCopyTaxonomy, deepCopyCategories, deepCopyEvents,
  deepCopyEventDef, deepCopySensitivity

Verification:
- go build ./... → clean
- go test -race -count=1 . → green (184s)

Refs #540.
